requirements.txt
google_auth_oauthlib==1.2.0
gspread==6.1.2
streamlit==1.36.0
python-dotenv==1.0.1